T #31 Remove NirvanaWiki.getPageLines() low-level implementation.
Replace it with simple one but not optimized.
Revert old changes in Wiki.java for this feature.
It was made for speed optimization when your fetch large archive pages
with a lot of new lines. I had not measured time benefit before I did this
optimization. And now i think that it's too mean.
It is hard to support this feature because it produces conflicts
every time when you want to update Wiki.java.
So we say: it is not worth it.