[#3,#58] Write that large and heavy test for NewPagesWeek updatement.
In NewPagesWeek the special logic for removing
old header & footer works, and we should cover this logic
with tests for continuing our refactoring works of #58 with safety.
Also, we have never had tests for this type of portal page,
now we will have it - good for #3.

Additionally:
- Add new methods to DateTools that work properly with timestamps
and that do not have a spoiled timezone.
- Start using new SystemTime wrapper class for easy mocking
time/date related code.
- Fix random order of updates of pages of the week by using
LinkedHashMap instead of HashMap.
- Add Python script for easy mocking large petscan responces.